# openapi/admin.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ultrafast Gateway Admin API",
    "description": "The admin API of the Ultrafast gateway, served under /api.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "2.0.0-alpha.1"
  },
  "paths": {
...
    "/api/models": {
      "get": {
        "tags": [
          "models"
        ],
        "operationId": "models_list",
        "responses": {
          "200": {
            "description": "An admin gets every model with its grants. Everyone else gets the enabled models they may call, with empty grants.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelList"
                }
              }
            }
          },
          "401": {
            "description": "No valid session or access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "token": []
          }
        ]
      },
      "post": {
        "tags": [
          "models"
        ],
        "operationId": "models_create",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "description": "The CSRF token of the session. Required with a session cookie; not needed with an access token.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The new model. It is disabled and granted to nobody.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelView"
                }
              }
            }
          },
          "400": {
            "description": "The request is not of the expected form.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid session or access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "The caller is not allowed to do this, or the CSRF token is missing or does not match.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "`model_exists`: the provider already has a model of this name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "The request body is too large.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Some fields are not valid; `fields` names each of them.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "token": []
          }
        ]
      }
    },
...
    "/api/models/{id}/grants": {
      "put": {
        "tags": [
          "models"
        ],
        "operationId": "models_grants_put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the model.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "description": "The CSRF token of the session. Required with a session cookie; not needed with an access token.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantsView"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The model with its new grants. They replace the old ones.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelView"
                }
              }
            }
          },
          "400": {
            "description": "The request is not of the expected form.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid session or access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "The caller is not allowed to do this, or the CSRF token is missing or does not match.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "It does not exist, or it is hidden from the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "The request body is too large.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "A team or user does not exist, or everyone is combined with others; `fields` names each of them.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "Every error of `/api` has this shape.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        }
      },
...
      "CreateModelRequest": {
        "type": "object",
        "required": [
          "provider_id",
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The provider's id for the model, 1 to 200 characters, no whitespace."
          },
          "provider_id": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "ModelList": {
        "type": "object",
        "required": [
          "models"
        ],
        "properties": {
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelView"
            }
          }
        }
      },
      "ModelView": {
        "type": "object",
        "required": [
          "id",
          "provider_id",
          "provider_name",
          "name",
          "enabled",
          "grants",
          "created_at",
          "input_price_micros",
          "output_price_micros"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "grants": {
            "$ref": "#/components/schemas/GrantsView"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "input_price_micros": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "What a million input tokens cost, in millionths of a dollar. `null`\nis unknown: calls of the model are logged with cost 0, unpriced."
          },
          "name": {
            "type": "string",
            "description": "The provider's own id for the model."
          },
          "output_price_micros": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "What a million output tokens cost, in millionths of a dollar."
          },
          "provider_id": {
            "type": "integer",
            "format": "int64"
          },
          "provider_name": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "uf_session",
        "description": "The session cookie set by signing in."
      },
      "token": {
        "type": "http",
        "scheme": "bearer",
        "description": "An access token, which starts with `uf-at-`."
      }
    }
  },
...
}